+++ CmdLine.hh
@@ -419,9 +419,12 @@ class CmdLine {
   /// @param prefix is the string the precedes each description line (default is "# ")
   /// @param absence_prefix is the string that precedes each line for an option that was not present
   /// @param presence_prefix is the string that precedes each line for an option that was present
+  /// @param compact if true, the output is more compact (no help lines)
   std::string dump(const std::string & prefix = "# ", 
                    const std::string & absence_prefix = "// ",
-                   const std::string & presence_prefix = "") const;
+                   const std::string & presence_prefix = "",
+                   bool compact = false
+                  ) const;
   
   /// return true if all options have been asked for at some point or other
   /// and send diagnostic info to ostr